moved spool into its own utility
